+++ browser/components/Game/Character/_index.scss
@@ -1,0 +1,55 @@
+@import "./Card/_index";
+
+
+#game-character-card-wrapper {
+  flex-grow: 0;
+  > div {
+    position: relative;
+    border: 1px solid black;
+    border-radius: 20px 20px 0 0;
+  }
+}
+#game-character-card-2 {
+  background-color: #7B68EE;
+}
+
+#game-character-card-3 {
+  background-color: #3CB371;
+}
+
+#game-character-card-4 {
+  background-color: #DAA520;
+}
+
+#game-board-error {
+  position: absolute;
+  top: 0;
+  left: 0;
+  bottom: 0;
+  right: 0;
+  z-index: 10;
+  background-color: rgba(201, 201, 201, 0.6);
+
+  div {
+    position: absolute;
+    display: inline-block;
+    height: auto !important;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    font-weight: bold;
+    font-size: 300%;
+  }
+}
+
+#game-board-status-overlay {
+  position: absolute !important;
+  top: 0;
+  right: 0;
+  width: 100%;
+  height: 100%;
+
+  .game-board-grid {
+    padding-right: 10px;
+  }
+}
