+++ browser/components/Game/_index.scss
@@ -1,5 +1,5 @@
 @import './_MonsterInfo';
-@import './Character/_Card';
+@import './Character/_index';
 
 #game {
 	flex-direction: column;
